feat: upgrade agent to latest version
