+++ agent/package.json
@@ -6,143 +6,23 @@
     "scripts": {
         "start": "node --loader ts-node/esm src/index.ts",
         "dev": "node --loader ts-node/esm src/index.ts",
-        "check-types": "tsc --noEmit"
+        "check-types": "tsc --noEmit",
+        "test": "jest"
+    },
+    "nodemonConfig": {
+        "watch": [
+            "src",
+            "../core/dist"
+        ],
+        "ext": "ts,json",
+        "exec": "node --enable-source-maps --loader ts-node/esm src/index.ts"
     },
     "dependencies": {
-        "@elizaos/adapter-supabase": "workspace:*",
-        "@elizaos/adapter-pglite": "workspace:*",
-        "@elizaos/adapter-postgres": "workspace:*",
-        "@elizaos/adapter-redis": "workspace:*",
-        "@elizaos/adapter-sqlite": "workspace:*",
-        "@elizaos/adapter-qdrant": "workspace:*",
-        "@elizaos/adapter-mongodb": "workspace:*",
-        "@elizaos/client-auto": "workspace:*",
         "@elizaos/client-direct": "workspace:*",
-        "@elizaos/client-discord": "workspace:*",
-        "@elizaos/client-farcaster": "workspace:*",
-        "@elizaos/client-lens": "workspace:*",
-        "@elizaos/client-telegram": "workspace:*",
-        "@elizaos/client-telegram-account": "workspace:*",
-        "@elizaos/client-twitter": "workspace:*",
-        "@elizaos/client-instagram": "workspace:*",
-        "@elizaos/client-slack": "workspace:*",
-        "@elizaos/client-alexa": "workspace:*",
-        "@elizaos/client-simsai": "workspace:*",
         "@elizaos/core": "workspace:*",
-        "@elizaos/plugin-0g": "workspace:*",
-        "@elizaos/plugin-abstract": "workspace:*",
-        "@elizaos/plugin-agentkit": "workspace:*",
-        "@elizaos/plugin-aptos": "workspace:*",
-        "@elizaos/plugin-birdeye": "workspace:*",
-        "@elizaos/plugin-zerion": "workspace:*",
-        "@elizaos/plugin-binance": "workspace:*",
-        "@elizaos/plugin-avail": "workspace:*",
-        "@elizaos/plugin-bnb": "workspace:*",
         "@elizaos/plugin-bootstrap": "workspace:*",
-        "@elizaos/plugin-di": "workspace:*",
-        "@elizaos/plugin-intiface": "workspace:*",
-        "@elizaos/plugin-coinbase": "workspace:*",
-        "@elizaos/plugin-coingecko": "workspace:*",
-        "@elizaos/plugin-coinmarketcap": "workspace:*",
-        "@elizaos/plugin-conflux": "workspace:*",
-        "@elizaos/plugin-cosmos": "workspace:*",
-        "@elizaos/plugin-desk-exchange": "workspace:*",
-        "@elizaos/plugin-echochambers": "workspace:*",
-        "@elizaos/plugin-evm": "workspace:*",
-        "@elizaos/plugin-edwin": "workspace:*",
-        "@elizaos/plugin-flow": "workspace:*",
-        "@elizaos/plugin-gelato": "workspace:*",
-        "@elizaos/plugin-giphy": "workspace:*",
-        "@elizaos/plugin-gitbook": "workspace:*",
-        "@elizaos/plugin-gitcoin-passport": "workspace:*",
-        "@elizaos/plugin-goat": "workspace:*",
-        "@elizaos/plugin-icp": "workspace:*",
-        "@elizaos/plugin-initia": "workspace:*",
-        "@elizaos/plugin-image-generation": "workspace:*",
-        "@elizaos/plugin-lens-network": "workspace:*",
-        "@elizaos/plugin-lit": "workspace:*",
-        "@elizaos/plugin-massa": "workspace:*",
-        "@elizaos/plugin-mind-network": "workspace:*",
-        "@elizaos/plugin-moralis": "workspace:*",
-        "@elizaos/plugin-movement": "workspace:*",
-        "@elizaos/plugin-news": "workspace:*",
-        "@elizaos/plugin-nft-generation": "workspace:*",
-        "@elizaos/plugin-node": "workspace:*",
-        "@elizaos/plugin-quick-intel": "workspace:*",
-        "@elizaos/plugin-solana": "workspace:*",
-        "@elizaos/plugin-solana-v2": "workspace:*",
-        "@elizaos/plugin-injective": "workspace:*",
-        "@elizaos/plugin-solana-agent-kit": "workspace:*",
-        "@elizaos/plugin-squid-router": "workspace:*",
-        "@elizaos/plugin-autonome": "workspace:*",
-        "@elizaos/plugin-starknet": "workspace:*",
-        "@elizaos/plugin-story": "workspace:*",
-        "@elizaos/plugin-sui": "workspace:*",
-        "@elizaos/plugin-sgx": "workspace:*",
-        "@elizaos/plugin-iq6900": "workspace:*",
-        "@elizaos/plugin-tee": "workspace:*",
-        "@elizaos/plugin-tee-log": "workspace:*",
-        "@elizaos/plugin-tee-marlin": "workspace:*",
-        "@elizaos/plugin-ton": "workspace:*",
-        "@elizaos/plugin-twitter": "workspace:*",
-        "@elizaos/plugin-primus": "workspace:*",
-        "@elizaos/plugin-cronoszkevm": "workspace:*",
-        "@elizaos/plugin-cronos": "workspace:*",
-        "@elizaos/plugin-3d-generation": "workspace:*",
-        "@elizaos/plugin-fuel": "workspace:*",
-        "@elizaos/plugin-avalanche": "workspace:*",
-        "@elizaos/plugin-video-generation": "workspace:*",
-        "@elizaos/plugin-web-search": "workspace:*",
-        "@elizaos/plugin-dexscreener": "workspace:*",
-        "@elizaos/plugin-letzai": "workspace:*",
-        "@elizaos/plugin-thirdweb": "workspace:*",
-        "@elizaos/plugin-genlayer": "workspace:*",
-        "@elizaos/plugin-tee-verifiable-log": "workspace:*",
-        "@elizaos/plugin-depin": "workspace:*",
-        "@elizaos/plugin-open-weather": "workspace:*",
-        "@elizaos/plugin-obsidian": "workspace:*",
-        "@elizaos/plugin-arthera": "workspace:*",
-        "@elizaos/plugin-allora": "workspace:*",
-        "@elizaos/plugin-opacity": "workspace:*",
-        "@elizaos/plugin-hyperliquid": "workspace:*",
-        "@elizaos/plugin-akash": "workspace:*",
-        "@elizaos/plugin-quai": "workspace:*",
-        "@elizaos/plugin-lightning": "workspace:*",
-        "@elizaos/plugin-b2": "workspace:*",
-        "@elizaos/plugin-nft-collections": "workspace:*",
-        "@elizaos/plugin-pyth-data": "workspace:*",
-        "@elizaos/plugin-openai": "workspace:*",
-        "@elizaos/plugin-devin": "workspace:*",
-        "@elizaos/plugin-holdstation": "workspace:*",
-        "@elizaos/plugin-router-nitro": "workspace:*",
-        "@elizaos/plugin-nvidia-nim": "workspace:*",
-        "@elizaos/plugin-0x": "workspace:*",
-        "@elizaos/plugin-bittensor": "workspace:*",
-        "@elizaos/plugin-chainbase": "workspace:*",
-        "@elizaos/plugin-dkg": "workspace:*",
-        "@elizaos/plugin-email": "workspace:*",
-        "@elizaos/plugin-sei": "workspace:*",
-        "@elizaos/plugin-omniflix": "workspace:*",
-        "@elizaos/plugin-suno": "workspace:*",
-        "@elizaos/plugin-udio": "workspace:*",
-        "@elizaos/plugin-hyperbolic": "workspace:*",
-        "@elizaos/plugin-football": "workspace:*",
-        "@elizaos/plugin-imgflip": "workspace:*",
-        "@elizaos/plugin-ethstorage": "workspace:*",
-        "@elizaos/plugin-mina": "workspace:*",
-        "@elizaos/plugin-email-automation": "workspace:*",
-        "@elizaos/plugin-dcap": "workspace:*",
-        "@elizaos/plugin-form": "workspace:*",
-        "@elizaos/plugin-ankr": "workspace:*",
-        "@elizaos/client-xmtp": "workspace:*",
-        "@elizaos/plugin-trikon": "workspace:*",
-        "@elizaos/plugin-zilliqa": "workspace:*",
-        "@elizaos/client-deva": "workspace:*",
-        "@elizaos/plugin-arbitrage": "workspace:*",
-        "@elizaos/plugin-multiversx": "workspace:*",
-        "@elizaos/plugin-near": "workspace:*",
-        "@elizaos/plugin-stargaze": "workspace:*",
-        "@elizaos/plugin-zksync-era": "workspace:*",
+        "@elizaos-plugins/adapter-sqlite": "github:elizaos-plugins/adapter-sqlite",
+        "@elizaos-plugins/plugin-di": "workspace:*",
         "@fixes-ai/core": "workspace:*",
         "@fixes-ai/common": "workspace:*",
         "readline": "1.3.0",
